1) Youtube Video Summarizer Project/requirements.txt
protobuf==3.20.3
python-dotenv==1.1.0
streamlit==1.42.2
youtube_transcript_api==1.0.3
google-generativeai==0.3.2